+++ datacrawler/acceptance_tests/import_vigie_rh_profession_groupe_test.py
@@ -113,7 +113,6 @@ class TestImportVigiePprofessionGroupe:
             inserer_nouvelles_donnees(
                 TABLE_REF_PROFESSION_GROUPE,
                 connection,
-                base_de_données_test,
                 SOURCE,
                 data_frames[0],
                 mocked_logger,
@@ -123,7 +122,6 @@ class TestImportVigiePprofessionGroupe:
             inserer_nouvelles_donnees(
                 TABLE_PROFESSION_GROUPE,
                 connection,
-                base_de_données_test,
                 SOURCE,
                 data_frames[1],
                 mocked_logger,